+++ src/variables/TarteaucitronVariable.php
@@ -4,6 +4,7 @@ namespace lhs\tarteaucitron\variables;
 
 use lhs\tarteaucitron\Tarteaucitron as Plugin;
 use Twig\Markup;
+use yii\base\InvalidConfigException;
 
 /**
  * Class TarteaucitronVariable
@@ -13,6 +14,7 @@ class TarteaucitronVariable
 {
     /**
      * @return Markup
+     * @noinspection PhpUnused Public API
      */
     public function initScript(): Markup
     {
@@ -22,14 +24,17 @@ class TarteaucitronVariable
     /**
      * @param array $options See [[HTML::tag()]] for details on how options are being used.
      * @return Markup
+     * @throws InvalidConfigException
+     * @noinspection PhpUnused Public API
      */
-    public function javascriptImportTag($options = []): Markup
+    public function javascriptImportTag(array $options = []): Markup
     {
         return Plugin::getInstance()->tarteaucitron->renderJavascriptImportTag($options);
     }
 
     /**
      * @return Markup
+     * @noinspection PhpUnused Public API
      */
     public function javascriptConfigTag(): Markup
     {
@@ -39,8 +44,10 @@ class TarteaucitronVariable
     /**
      * @param array $options See [[HTML::tag()]] for details on how options are being used.
      * @return Markup
+     * @throws InvalidConfigException
+     * @noinspection PhpUnused Public API
      */
-    public function stylesheetTag($options = []): Markup
+    public function stylesheetTag(array $options = []): Markup
     {
         return Plugin::getInstance()->tarteaucitron->renderStylesheetTag($options);
     }
@@ -48,6 +55,7 @@ class TarteaucitronVariable
     /**
      * @param array $options
      * @return Markup
+     * @noinspection PhpUnused Public API
      */
     public function reCAPTCHA(array $options = []): Markup
     {
@@ -57,6 +65,7 @@ class TarteaucitronVariable
     /**
      * @param array $options
      * @return Markup
+     * @noinspection PhpUnused Public API
      */
     public function googleMaps(array $options = []): Markup
     {
@@ -66,6 +75,7 @@ class TarteaucitronVariable
     /**
      * @param array $options
      * @return Markup
+     * @noinspection PhpUnused Public API
      */
     public function googleAdWordsConversion(array $options = []): Markup
     {
@@ -75,6 +85,7 @@ class TarteaucitronVariable
     /**
      * @param array $options
      * @return Markup
+     * @noinspection PhpUnused Public API
      */
     public function linkedin(array $options = []): Markup
     {
@@ -84,6 +95,7 @@ class TarteaucitronVariable
     /**
      * @param array $options
      * @return Markup
+     * @noinspection PhpUnused Public API
      */
     public function twitterFollowButton(array $options = []): Markup
     {
@@ -93,6 +105,7 @@ class TarteaucitronVariable
     /**
      * @param array $options
      * @return Markup
+     * @noinspection PhpUnused Public API
      */
     public function twitterShareButton(array $options = []): Markup
     {
@@ -102,6 +115,7 @@ class TarteaucitronVariable
     /**
      * @param array $options
      * @return Markup
+     * @noinspection PhpUnused Public API
      */
     public function vimeo(array $options = []): Markup
     {
@@ -111,67 +125,116 @@ class TarteaucitronVariable
     /**
      * @param array $options
      * @return Markup
+     * @noinspection PhpUnused Public API
      */
     public function youtube(array $options = []): Markup
     {
         return Plugin::getInstance()->tarteaucitron->renderYoutube($options);
     }
 
+    /**
+     * @return bool
+     * @noinspection PhpUnused Public API
+     */
     public function isFacebookPixelEnabled(): bool
     {
         return Plugin::getInstance()->tarteaucitron->isFacebookPixelEnabled();
     }
 
+    /**
+     * @return bool
+     * @noinspection PhpUnused Public API
+     */
     public function isGoogleTagManagerEnabled(): bool
     {
         return Plugin::getInstance()->tarteaucitron->isGoogleTagManagerEnabled();
     }
 
+    /**
+     * @return bool
+     * @noinspection PhpUnused Public API
+     */
     public function isReCAPTCHAEnabled(): bool
     {
         return Plugin::getInstance()->tarteaucitron->isReCAPTCHAEnabled();
     }
 
+    /**
+     * @return bool
+     * @noinspection PhpUnused Public API
+     */
     public function isGoogleMapsEnabled(): bool
     {
         return Plugin::getInstance()->tarteaucitron->isGoogleMapsEnabled();
     }
 
+    /**
+     * @return bool
+     * @noinspection PhpUnused Public API
+     */
     public function isGoogleAnalyticsUniversalEnabled(): bool
     {
         return Plugin::getInstance()->tarteaucitron->isGoogleAnalyticsUniversalEnabled();
     }
 
+    /**
+     * @return bool
+     * @noinspection PhpUnused Public API
+     */
     public function isGoogleAdWordsConversionEnabled(): bool
     {
         return Plugin::getInstance()->tarteaucitron->isGoogleAdWordsConversionEnabled();
     }
 
+    /**
+     * @return bool
+     * @noinspection PhpUnused Public API
+     */
     public function isGoogleAdWordsRemarketingEnabled(): bool
     {
         return Plugin::getInstance()->tarteaucitron->isGoogleAdWordsRemarketingEnabled();
     }
 
+    /**
+     * @return bool
+     * @noinspection PhpUnused Public API
+     */
     public function isLinkedInEnabled(): bool
     {
         return Plugin::getInstance()->tarteaucitron->isLinkedInEnabled();
     }
 
+    /**
+     * @return bool
+     * @noinspection PhpUnused Public API
+     */
     public function isTwitterEnabled(): bool
     {
         return Plugin::getInstance()->tarteaucitron->isTwitterEnabled();
     }
 
+    /**
+     * @return bool
+     * @noinspection PhpUnused Public API
+     */
     public function isVimeoEnabled(): bool
     {
         return Plugin::getInstance()->tarteaucitron->isVimeoEnabled();
     }
 
+    /**
+     * @return bool
+     * @noinspection PhpUnused Public API
+     */
     public function isYoutubeEnabled(): bool
     {
         return Plugin::getInstance()->tarteaucitron->isYoutubeEnabled();
     }
 
+    /**
+     * @return bool
+     * @noinspection PhpUnused Public API
+     */
     public function isYoutubeJsApiEnabled(): bool
     {
         return Plugin::getInstance()->tarteaucitron->isYoutubeJsApiEnabled();